qcacld-3.0: Fix undefined shift for DOT11BN in fw_wlan_feat_caps
g_fw_wlan_feat_caps is declared as uint32_t but enum cap_bitmap
now has DOT11BN = 32. Shifting 1 << 32 in wma_set_fw_wlan_feat_caps()
and wma_get_fw_wlan_feat_caps() is undefined behaviour since the
shift count equals the bit-width of int, so the 11BN capability bit
is not reliably set or read.

Widen g_fw_wlan_feat_caps to uint64_t and use 1ULL for the shift so
all 33 defined feature bits, including DOT11BN, can be represented
and accessed correctly.

CRs-Fixed: 4615669

+++ core/wma/src/wma_main.c
@@ -132,7 +132,7 @@
 #define WMA_LOG_COMPLETION_TIMER 500 /* 500 msecs */
 #define WMI_TLV_HEADROOM 128
 
-static uint32_t g_fw_wlan_feat_caps;
+static uint64_t g_fw_wlan_feat_caps;
 /**
  * wma_get_fw_wlan_feat_caps() - get fw feature capability
  * @feature: feature enum value
@@ -141,7 +141,7 @@ static uint32_t g_fw_wlan_feat_caps;
  */
 bool wma_get_fw_wlan_feat_caps(enum cap_bitmap feature)
 {
-	return (g_fw_wlan_feat_caps & (1 << feature)) ? true : false;
+	return (g_fw_wlan_feat_caps & (1ULL << feature)) ? true : false;
 }
 
 /**
@@ -152,7 +152,7 @@ bool wma_get_fw_wlan_feat_caps(enum cap_bitmap feature)
  */
 void wma_set_fw_wlan_feat_caps(enum cap_bitmap feature)
 {
-	g_fw_wlan_feat_caps |= (1 << feature);
+	g_fw_wlan_feat_caps |= (1ULL << feature);
 }
 
 /**
